pyproject.toml
[project]
name = "lang-graph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain-ollama>=1.0.1",
    "langchain[anthropic]>=1.2.13",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]
